% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_data_edges}
\alias{add_data_edges}
\title{Add data edges}
\usage{
add_data_edges(data, group_col_chr)
}
\arguments{
\item{data}{A data frame input from \code{new_ternable()}}

\item{group_col_chr}{Character vector of group column names}
}
\description{
Internal helper to create paths/edges between observations in a ternary plot.
Used by \code{new_ternable()} to create the \code{data_edges} component.
}
\keyword{internal}
