% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ternary_base.R
\name{add_ternary_base}
\alias{add_ternary_base}
\title{Draw the 2D ternary simplex}
\usage{
add_ternary_base(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:geom_polygon]{ggplot2::geom_polygon()}}, such as
\code{colour}, \code{fill}, \code{linewidth}, etc.}
}
\value{
A ggplot object
}
\description{
Draws the boundary of a 2D ternary simplex as an equilateral triangle
}
\examples{
library(ggplot2)

# Basic simplex
ggplot() + add_ternary_base()

# Customize appearance
ggplot() + add_ternary_base(colour = "blue", linewidth = 1.5)

}
