% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elb_map}
\alias{elb_map}
\title{Electoral boundaries map for the 2025 Australian Federal Election}
\format{
A tibble of 8 columns:
\describe{
\item{long}{Longitude of point in polygon}
\item{lat}{Latitude of point in polygon}
\item{hole}{Whether the polygon has a hole}
\item{piece}{Polygon piece number}
\item{group}{Polygon group number}
\item{order}{Order of polygon within group}
\item{id}{Unique identifier for polygon}
\item{elect_div}{Electoral division name}
}
}
\source{
Australian Electoral Commission (AEC)
\url{https://www.aec.gov.au/electorates/maps.htm}
}
\usage{
elb_map
}
\description{
Provides the points that make up the boundaries of each electoral division in the
2025 Australian Federal Election.
}
\examples{
library(ggplot2)
library(ggthemes)

# Load the dataset
data(elb_map)

# Plot the map
ggplot(elb_map) + 
  geom_polygon(
    aes(x = long, y = lat, group = group),
    fill = "grey90", color = "white") +
  theme_map()
}
\keyword{datasets}
