% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ternary_cart.R
\name{geom_ternary_cart}
\alias{geom_ternary_cart}
\title{Draw the 2D ternary simplex}
\usage{
geom_ternary_cart(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:geom_polygon]{ggplot2::geom_polygon()}}, such as
\code{colour}, \code{fill}, \code{linewidth}, etc.}
}
\description{
Draws the boundary of a 2D ternary simplex as an equilateral triangle
}
\examples{
library(ggplot2)

# Basic simplex
ggplot() + geom_ternary_cart()

# Customize appearance
ggplot() + geom_ternary_cart(colour = "blue", linewidth = 1.5)

}
