% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmert_trans.R
\name{ggtern_cart2d}
\alias{ggtern_cart2d}
\title{Ternary plot in cartesian coordinates}
\usage{
ggtern_cart2d(data, alternatives = NULL, label = TRUE, ...)
}
\description{
Ternary plot in cartesian coordinates
}
