% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefviz-package.R
\docType{package}
\name{prefviz-package}
\alias{prefviz}
\alias{prefviz-package}
\title{prefviz: Visualizes Preferential Data in One and More Contests}
\description{
A visualization toolkit for preferential data, such as ranked-choice election results, tournament outcomes, and survey responses. The package provides methods to visualise the preference distribution of one contest with bar charts and pairwise comparisons of two contestants, as well as methods to visualise multiple contests through 2D and high-dimensional simplex plots both statically and interactively. HD simplex displays are implemented via projection methods using the 'tourr' and 'detourr' packages, enabling dynamic exploration of high-dimensional preference structure. For more details on HD simplex projection, see Wickham et al. (2011) \doi{10.21105/joss.03419}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://numbats.github.io/prefviz/}
  \item Report bugs at \url{https://github.com/numbats/prefviz/issues}
}

}
\author{
\strong{Maintainer}: Linh Ngo \email{linhngo66.work@gmail.com}

Authors:
\itemize{
  \item Dianne Cook \email{DiCook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
  \item Damjan Vukcevic \email{damjan.vukcevic@monash.edu} (\href{https://orcid.org/0000-0001-7780-9586}{ORCID})
}

}
\keyword{internal}
