% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{print.pairwise}
\alias{print.pairwise}
\title{Print a pairwise object}
\usage{
\method{print}{pairwise}(x, ...)
}
\arguments{
\item{x}{A \code{pairwise} object returned by \code{\link[=pairwise_calculator]{pairwise_calculator()}}.}

\item{...}{Currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a pairwise object
}
\keyword{internal}
