% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternable.R
\name{print.ternable}
\alias{print.ternable}
\title{Print method for ternable objects}
\usage{
\method{print}{ternable}(x, ...)
}
\arguments{
\item{x}{A ternable object}

\item{...}{Additional arguments passed to print methods}
}
\value{
The object, invisibly
}
\description{
Print method for ternable objects
}
\keyword{internal}
