% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ternary_region.R
\docType{data}
\name{geom_ternary_region}
\alias{geom_ternary_region}
\alias{StatTernaryRegion}
\alias{stat_ternary_region}
\title{Create 3 polygonial regions in a ternary plot based on a reference point}
\format{
An object of class \code{StatTernaryRegion} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatTernaryRegion

stat_ternary_region(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  show.legend = NA,
  inherit.aes = FALSE,
  x1 = 1/3,
  x2 = 1/3,
  x3 = 1/3,
  vertex_labels = NULL,
  ...
)

geom_ternary_region(
  mapping = NULL,
  data = NULL,
  position = "identity",
  show.legend = NA,
  inherit.aes = TRUE,
  x1 = 1/3,
  x2 = 1/3,
  x3 = 1/3,
  vertex_labels = NULL,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}.
To map aesthetics to computed variables (like region labels), use
\code{\link[ggplot2:aes_eval]{ggplot2::after_stat()}}, e.g., \code{aes(fill = after_stat(vertex_labels))}.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data for \code{stat_ternary_region()}.
Default is \code{"polygon"}.}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{show.legend}{Logical. Determines whether this layer is included in the legends.
\code{NA} (the default) includes if any aesthetics are mapped. \code{FALSE} never includes,
and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than
combining with them. This is most useful for helper functions that define both
data and aesthetics and shouldn't inherit behaviour from the default plot specification.

#' @section Computed variables:
These are calculated by \code{stat_ternary_region()}:
\describe{
\item{\code{x}, \code{y}}{Cartesian coordinates of polygon vertices}
\item{\code{id}}{Numeric identifier for each points (
\itemize{
\item 1-3 are the vertices of the triangle
\item 4 is the reference point
\item 5-7 are the projection of the reference point to the edges of the triangle)}
\item{\code{group}}{Numeric identifier for each region (1, 2, or 3)}
\item{\code{vertex_labels}}{Label for the vertex, representing each region.
If \code{vertex_labels} is \code{NULL}, the labels default to \verb{Region 1, Region 2, Region 3}.}
}
}}

\item{x1, x2, x3}{Numeric values defining the reference point in ternary coordinates.
Must sum to 1 (or will be normalized). Default is \verb{1/3, 1/3, 1/3} (centroid),
diving the ternary space into 3 equal regions.}

\item{vertex_labels}{Character vector of length 3 giving labels for the three
regions. The order corresponds to the three vertices of the ternary triangle
(see Details). If \code{NULL}, regions are labeled numerically. These labels can be
accessed in aesthetic mappings via \code{after_stat(vertex_labels)}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}
}
\description{
\code{geom_ternary_region()} and \code{stat_ternary_region()} create three polygonal
regions in a ternary plot via a reference point. Each region is defined by perpendicular
projections from the reference point to the edges of the ternary triangle.
Each region represents one of the traingular vertex that forms the corresponding region.
}
\section{Aesthetic mappings}{


geom_ternary_region()\verb{uses [ggplot2::geom_polygon()], so it understands the  same aesthetics. The most commonly used are: \\itemize\{ \\item}fill\verb{- Fill color of regions \\item}colour\code{/}color\verb{- Border color of regions \\item}alpha\verb{- Transparency (0 = transparent, 1 = opaque) \\item}linewidth\verb{- Width of region borders \\item}linetype` - Type of border lines
}

However, since this geom generates its own data based on the reference point parameters, \verb{ }after_stat()\verb{is needed to access the computed variables.  The most common use case is to map aesthetics based on the three alternatives,  representing by three ternary vertices.  To do so, use}ggplot2::after_stat(vertex_labels)\verb{in your}aes()\verb{mappings.  See Computed variables section for variables calculated by}stat_ternary_region()`.

TODO: Update with examples once geom_ternary_cart() is done
}

\keyword{datasets}
