% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefviz-package.R
\docType{package}
\name{prefviz-package}
\alias{prefviz}
\alias{prefviz-package}
\title{prefviz: Visualizes preference data via ternary plots in 2 and higher dimensions}
\description{
Visualises preference and ranking data by extending traditional ternary plots to support high-dimensional simplexes. It provides methods to transform compositional data into coordinates suitable for 2D and high-dimensional ternary plots. Compatibility with interacitve visualization packages such as `plotly` allows users to explore high-dimensional preference structures dynamically.
}
\author{
\strong{Maintainer}: Linh Ngo \email{Linh.Ngo2@monash.edu}

Other contributors:
\itemize{
  \item Dianne Cook \email{DiCook@monash.edu} (Supervisor) [contributor]
  \item Damjan Dvukcevic \email{Damjan@Vukcevic.net} (Supervisor) [contributor]
}

}
\keyword{internal}
